+++ docs/changelog.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Flag for ScLogger to enable appending logs to existing log file
 
+### Fixed 
+
+- Context template argument initialization in ScAgent class 
+
 ## [0.10.3] - 01.05.2025
 
 ### Fixed