+++ docs/changelog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `GetElements` method for `ScSet` class to get all elements of set with their roles
+- `GetElementsByRoles` method for `ScSet` class to get all elements of set by their roles
 - `GetElements` method for `ScSet` class to get all elements of set
 - ScOrientedSet class to work with sets of sc-elements ordered by nrel_basic_sequence relation
 - Methods for `ScSet` class: `Next`, `Reset`, `ForEach`